Split tests to various targets
