+++ tests/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.22)
-
 project(senders-io_tests CXX)
-include(CTest)
+
+set(SIO_TESTS_ROOT_DIR ${CMAKE_CURRENT_SOURCE_DIR})
 
 if (PROJECT_IS_TOP_LEVEL)
   find_package(sio REQUIRED)
@@ -28,7 +28,7 @@ set(SIO_TEST_DEFINES
 
 function(sio_add_test_target target_name)
   add_executable(${target_name} ${ARGN})
-  target_include_directories(${target_name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+  target_include_directories(${target_name} PRIVATE ${SIO_TESTS_ROOT_DIR} ${CMAKE_CURRENT_SOURCE_DIR})
   target_link_libraries(${target_name}
     PRIVATE
       sio::sio
@@ -38,40 +38,17 @@ function(sio_add_test_target target_name)
   add_test(NAME ${target_name} COMMAND ${target_name})
 endfunction()
 
-set(SIO_TEST_FILE_SOURCES
-  file/test_file_handle.cpp
-  file/test_read_batched.cpp
-)
-
-set(SIO_TEST_NET_SOURCES
-  net/test_async_accept.cpp
-  net/test_can_endpoint.cpp
-  net/test_can_socket.cpp
-  net/test_address.cpp
-  net/test_endpoint.cpp
-  net/test_resolve.cpp
-  net/test_socket_handle.cpp
-)
-
 set(SIO_TEST_LIB_SOURCES
-  sequence/test_first.cpp
-  sequence/test_last.cpp
-  sequence/test_fork.cpp
-  sequence/test_merge_each.cpp
-  sequence/test_scan.cpp
-  sequence/test_transform_each.cpp
-  sequence/test_repeat.cpp
-  sequence/test_zip.cpp
-  sequence/test_finally.cpp
-  sequence/test_buffered_sequence.cpp
   # test_async_channel.cpp
-  test_async_resource.cpp
   test_async_mutex.cpp
+  test_async_resource.cpp
   test_const_buffer_subspan.cpp
   test_memory_pool.cpp
   test_tap.cpp
 )
 
-sio_add_test_target(sio_tests_file ${SIO_TEST_FILE_SOURCES})
 sio_add_test_target(sio_tests_lib ${SIO_TEST_LIB_SOURCES})
-sio_add_test_target(sio_tests_net ${SIO_TEST_NET_SOURCES})
+
+add_subdirectory(file)
+add_subdirectory(net)
+add_subdirectory(sequence)
