+++ tests/file/test_read_batched.cpp
@@ -5,7 +5,7 @@
 #include <sio/sequence/iterate.hpp>
 #include <sio/sequence/ignore_all.hpp>
 #include <sio/buffer.hpp>
-#include "backend_matrix.hpp"
+#include "common/backend_matrix.hpp"
 
 #include <catch2/catch_all.hpp>
