+++ tests/net/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(SIO_TEST_NET_SOURCES
+  test_async_accept.cpp
+  test_can_endpoint.cpp
+  test_can_socket.cpp
+  test_address.cpp
+  test_endpoint.cpp
+  test_resolve.cpp
+  test_socket_handle.cpp
+)
+
+sio_add_test_target(sio_tests_net ${SIO_TEST_NET_SOURCES})
