+++ tests/net/test_can_socket.cpp
@@ -1,4 +1,4 @@
-#include "backend_matrix.hpp"
+#include "common/backend_matrix.hpp"
 
 #include "sio/async_resource.hpp"
 #include "sio/can/raw_protocol.hpp"
