+++ tests/net/test_socket_handle.cpp
@@ -2,7 +2,7 @@
 #include "sio/event_loop/concepts.hpp"
 #include "sio/event_loop/socket_handle.hpp"
 #include "sio/ip/tcp.hpp"
-#include "backend_matrix.hpp"
+#include "common/backend_matrix.hpp"
 
 #include <catch2/catch_all.hpp>
